feat: Allow value view queries to return nothing.
Allows a QueryHandlerReturnsValue to throw a NotFound exception if it
does not have anything to return.

This includes a major rework of the execution of query handlers. It
splits the API endpoint for view querying into one for value queries and
one for stream queries, to make the distinction between the two clear
and allow different behavior.

This also includes a change of the handling of returned items with
mismatching schema. These will now be dropped in a way clients will not
notice by either omitting them from the result stream in the case of
stream queries or by returning a 404 status code in the case of value
queries.