chore: Integrate ContextIdentifier and AggregateIdentifier.
Since the ContextIdentifier was literally never used on its own and in
the few places where the AggregateIdentifier was used on its own we
really needed the context information, these two are now one. This means
that the DomainEventStore now also contains the context name of each
stored domain event.